refactor(cli): address PR #225 review warnings
Extract duplicated daemon test helpers into a shared utility and align note project-name resolution behavior with project/task command semantics for multiple matches.

Task: #1939
